# openapi/Skoleoverblikket.Api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Skoleoverblikket API",
    "version": "v1"
  },
  "paths": {
    "/api/v1/absence": {
      "post": {
        "tags": [
          "Absence"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbsenceController_ReportAbsenceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AbsenceController_ReportAbsenceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AbsenceController_ReportAbsenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Absence"
        ],
        "parameters": [
          {
            "name": "classId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AbsenceController_AbsenceReportDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AbsenceController_AbsenceReportDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AbsenceController_AbsenceReportDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/absence/{id}/confirm": {
      "post": {
        "tags": [
          "Absence"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/absence/{id}/dismiss": {
      "post": {
        "tags": [
          "Absence"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/absence/{id}": {
      "delete": {
        "tags": [
          "Absence"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/tenants/{schoolId}/modules": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "schoolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminController_ModuleOverrideRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminController_ModuleOverrideRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdminController_ModuleOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/tenants/{schoolId}/modules/{module}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "schoolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "module",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SubscriptionModule"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbsenceController_AbsenceReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "studentName": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/AbsenceStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "AbsenceController_ReportAbsenceRequest": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AbsenceStatus": {
        "enum": [
          "Reported",
          "Confirmed",
          "Dismissed"
        ],
        "type": "string"
      },
      "AdminController_ModuleOverrideRequest": {
        "type": "object",
        "properties": {
          "module": {
            "$ref": "#/components/schemas/SubscriptionModule"
          }
        },
        "additionalProperties": false
      },
...
      "SubscriptionModule": {
        "enum": [
          "ParentModule",
          "BoardModule"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}